pyproject.toml
[project]
name = "llm-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "jinja2>=3.1.6",
    "mmo-lr-utils",
    "numpy>=1.26.4",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_judge*", "prompt*", "response*", "utils*"]

[tool.uv.sources]
mmo-lr-utils = { path = "wheelhouse/mmo_lr_utils-1.0.1-py3-none-any.whl" }

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.30.1",
]
